:recycle: refactor: remodelando a aplicação seguindo os princípios de DDD (Domain Driven Design)     - Restruturando o projeto para abstrair e aplicar o conteito de DDD, e assim proporcionam maior organização das responsabilidades do código.     - Camada de aplicação (application): onde os controladores e serviços da API serão criados.     - Camada de dominio (domain): será o coração do projeto, onde serão implementadas as classes e interfaces do sistema que atendem as regras de negócio da aplicação.     - Camada de infraestrutura (infrastructure): quer será responsável por prover a camada de domínio os insumos necessários para que ela consiga processar as regras de negócio.
