+++ src/main/java/br/com/alelo/consumer/consumerpat/infrastructure/respository/ExtractRepository.java
@@ -1,6 +1,6 @@
-package br.com.alelo.consumer.consumerpat.respository;
+package br.com.alelo.consumer.consumerpat.infrastructure.respository;
 
-import br.com.alelo.consumer.consumerpat.entity.Extract;
+import br.com.alelo.consumer.consumerpat.domain.entity.Extract;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ExtractRepository extends JpaRepository<Extract, Integer> {
